Moving replayer out of lua
